feat: add and apply redis caching

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const AuthorizationError = require("../../exceptions/AuthorizationError");
 const { mapDBToModel } = require("../../utils/mapDBtoModel");
 
 class NotesService {
-    constructor(collaborationService) {
+    constructor(collaborationService, cacheService) {
         this._pool = new Pool();
         this._collaborationService = collaborationService;
+        this._cacheService = cacheService;
     }
 
     async addNote({ title, body, tags, owner }) {
@@ -26,19 +27,33 @@ class NotesService {
             throw new InvariantError("Catatan gagal ditambahkan");
         }
 
+        await this._cacheService.delete(`notes:${owner}`);
         return rows[0].id;
     }
 
     async getNotes(owner) {
-        const query = {
-            text: `SELECT notes.* FROM notes
-          LEFT JOIN collaborations ON collaborations.note_id = notes.id
-          WHERE notes.owner = $1 OR collaborations.user_id = $1
-          GROUP BY notes.id`,
-            values: [owner],
-        };
-        const { rows } = await this._pool.query(query);
-        return rows.map(mapDBToModel);
+        try {
+            const result = await this._cacheService.get(`notes:${owner}`);
+            return JSON.parse(result);
+        } catch (error) {
+            const query = {
+                text: `SELECT notes.* FROM notes
+              LEFT JOIN collaborations ON collaborations.note_id = notes.id
+              WHERE notes.owner = $1 OR collaborations.user_id = $1
+              GROUP BY notes.id`,
+                values: [owner],
+            };
+
+            const result = await this._pool.query(query);
+            const mappedResult = result.rows.map(mapDBToModel);
+
+            await this._cacheService.set(
+                `notes:${owner}`,
+                JSON.stringify(mappedResult)
+            );
+
+            return mappedResult;
+        }
     }
 
     async getNoteById(id) {
@@ -60,32 +75,38 @@ class NotesService {
     async editNoteById(id, { title, body, tags }) {
         const updatedAt = new Date().toISOString();
         const query = {
-            text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id",
+            text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner",
             values: [title, body, tags, updatedAt, id],
         };
 
-        const { rowCount } = await this._pool.query(query);
+        const { rows, rowCount } = await this._pool.query(query);
 
         if (!rowCount) {
             throw new NotFoundError(
                 "Gagal memperbarui catatan. Id tidak ditemukan"
             );
         }
+
+        const { owner } = rows[0];
+        await this._cacheService.delete(`notes:${owner}`);
     }
 
     async deleteNoteById(id) {
         const query = {
-            text: "DELETE FROM notes WHERE id = $1 RETURNING id",
+            text: "DELETE FROM notes WHERE id = $1 RETURNING id, owner",
             values: [id],
         };
 
-        const { rowCount } = await this._pool.query(query);
+        const { rows, rowCount } = await this._pool.query(query);
 
         if (!rowCount) {
             throw new NotFoundError(
                 "Catatan gagal dihapus. Id tidak ditemukan"
             );
         }
+
+        const { owner } = rows[0];
+        await this._cacheService.delete(`notes:${owner}`);
     }
 
     async verifyNoteOwner(id, owner) {
